+++ dopplerview/input_output/output_renderer.py
@@ -15,9 +15,18 @@ class SignalRenderer(OutputRenderer):
     def render(self, key, ctx, path, options=None):
         plt.figure()
         plt.title(key)
-        if options and options.get("multiple_signals"):
+
+        scatter_indices = None
+        if options and options.get("scatter"):
+            scatter_indices = ctx.get(key)[-1] # Only scatter the last signal if multiple_signals is True
+
+        if options and (options.get("multiple_signals") or options.get("scatter")):
             legend = options.get("legend", [])
             for i, signal in enumerate(ctx.get(key)):
+                if scatter_indices is not None:
+                    if i == len(ctx.get(key)) - 1:  # Skip las iteration if scatter is True
+                        continue
+                    plt.scatter(scatter_indices, signal[scatter_indices], label=f"{legend[i]} Peaks" if i < len(legend) else "Peaks", s=50)
                 plt.plot(signal, label=legend[i] if i < len(legend) else "")
             if legend:
                 plt.legend()
@@ -34,10 +43,10 @@ class OpticDiscRenderer(OutputRenderer):
     def render(self, key, ctx, path, options=None):
         image = ctx.get("M0_ff_image")
         center = ctx.get("optic_disc_center")
-        axes = ctx.get("optic_disc_axes")
+        diameter_x = ctx.get("optic_disc_width")
+        diameter_y = ctx.get("optic_disc_height")
 
         x_center, y_center = center
-        diameter_x, diameter_y = axes
 
         a = diameter_x / 2
         b = diameter_y / 2