+++ dopplerview/pipeline/steps/pulse_analysis.py
@@ -1,7 +1,9 @@
 from dopplerview.pipeline.step import BaseStep, NestedStep
 from dopplerview.segmentation import process_masks, pulse_analysis, signal_processing
 from dopplerview.utils.parallelization_utils import run_in_parallel
+import dopplerview.utils.image_utils as image_utils
 from functools import partial
+
 
 import numpy as np
 
@@ -17,7 +19,7 @@ class PulseAnalysisStep(NestedStep):
             
 class PreArteryMaskStep(BaseStep):
     requires = {"M0_ff_video", "retinal_vessel_mask", "optic_disc_center"}
-    produces = {"labeled_vessels", "pre_artery_mask", "branch_signals", "corrected_signals", "pre_vein_mask"}
+    produces = {"labeled_vessels", "pre_artery_mask", "branch_signals", "corrected_signals", "pre_vein_mask", "beat_period"}
     name = "pre_artery_mask"
 
     def _relevant_config(self, ctx):
@@ -50,12 +52,13 @@ class PreArteryMaskStep(BaseStep):
         ctx.output_manager.output("pulse_analysis", "labeled_vessels", labeled_vessels, "labeled_mask")
         signals_n = (signals - signals.mean(axis=1, keepdims=True)) / signals.std(axis=1, keepdims=True)
         ctx.set("branch_signals", signals_n)
+        ctx.output_manager.save_h5("branch_signals", ctx)
 
         # --- Step 3: Correct signals by aligning with median heartbeat ---
         beat_period = pulse_analysis.compute_idx0(signals_n, sampling_frequency)
         corrected_signals = np.zeros_like(signals_n)
-        func = partial(pulse_analysis.correct_branch_signal_with_heartbeat, beat_period=beat_period, k=10)
-        corrected_signals = run_in_parallel(func, signals_n, n_jobs=ctx.dopplerview_config["NumberOfWorkers"], chunking=False, task_name="signal correction")
+        func = partial(pulse_analysis.correct_branch_signal_with_heartbeat, beat_period=beat_period, k=5)
+        corrected_signals = run_in_parallel(func, signals_n, n_jobs=ctx.dopplerview_config["NumberOfWorkers"], chunking=False, task_name="branch signal correction")
         # for i, signal in enumerate(signals_n):
         #     corrected_signals[i, :] = pulse_analysis.correct_branch_signal_with_heartbeat(signal, beat_period, k=10)
         ctx.set("corrected_signals", corrected_signals)
@@ -67,10 +70,11 @@ class PreArteryMaskStep(BaseStep):
         pre_artery_mask, pre_vein_mask = pulse_analysis.compute_pre_masks(corrected_signals, labeled_vessels, sampling_frequency)
         ctx.set("pre_artery_mask", pre_artery_mask)
         ctx.set("pre_vein_mask", pre_vein_mask)
+        ctx.set("beat_period", beat_period)
 
 class ComputeTemporalCuesStep(BaseStep):
-    requires = {"M0_ff_video", "pre_artery_mask", "choroidal_vessel_mask"}
-    produces = {"correlation", "diasys_image", "pre_arterial_pulse", "choroidal_pulse", "pre_arterial_pulse_filtered", "choroidal_pulse_filtered", "pre_arterial_pulse_interpolated", "pre_venous_pulse", "pre_venous_pulse_filtered"}
+    requires = {"M0_ff_video", "pre_artery_mask", "choroidal_vessel_mask", "beat_period"}
+    produces = {"correlation", "diasys_image", "pre_arterial_pulse", "choroidal_pulse", "pre_arterial_pulse_filtered", "choroidal_pulse_filtered", "pre_arterial_pulse_cleaned", "pre_venous_pulse", "pre_venous_pulse_filtered", "M0_ff_image_cleaned"}
     name = "temporal_cues"
 
     def _relevant_config(self, ctx):
@@ -88,6 +92,9 @@ class ComputeTemporalCuesStep(BaseStep):
         arterial_pulse = signal_processing.get_pulse_from_mask(video, pre_artery_mask)
         venous_pulse = signal_processing.get_pulse_from_mask(video, pre_vein_mask)
         choroidal_pulse = signal_processing.get_pulse_from_mask(video, choroidal_vessel_mask)
+        ctx.set("pre_arterial_pulse", arterial_pulse)
+        ctx.set("pre_venous_pulse", venous_pulse)
+        ctx.set("choroidal_pulse", choroidal_pulse)
 
         # --- Filter pulses to remove high frequency noise ---
 
@@ -99,28 +106,36 @@ class ComputeTemporalCuesStep(BaseStep):
         arterial_pulse_filtered = signal_processing.get_filtered_pulse(arterial_pulse, sampling_frequency)
         venous_pulse_filtered = signal_processing.get_filtered_pulse(venous_pulse, sampling_frequency)
         choroidal_pulse_filtered = signal_processing.get_filtered_pulse(choroidal_pulse, sampling_frequency)
+        ctx.set("pre_arterial_pulse_filtered", arterial_pulse_filtered)
+        ctx.set("pre_venous_pulse_filtered", venous_pulse_filtered)
+        ctx.set("choroidal_pulse_filtered", choroidal_pulse_filtered)
 
-        # --- Interpolate outlier frames using the filtered signal ---
+        # --- Remove bad beats from arterial pulse by comparing to median beat pattern ---
 
-        video_cleaned, arterial_pulse_interpolated = signal_processing.interpolate_outliers(video, arterial_pulse, pre_artery_mask, sampling_frequency=sampling_frequency)
-        # ctx.output_manager.output("pulse_analysis", "video_cleaned", video_cleaned, "video")
+        beat_period = ctx.require("beat_period")
+
+        arterial_pulse_cleaned, video_cleaned, beat_signal = pulse_analysis.remove_bad_beats(arterial_pulse_filtered, video, beat_period, threshold=0.8)
+        ctx.set("pre_arterial_pulse_cleaned", arterial_pulse_cleaned)
+
+        M0_ff_image_cleaned = image_utils.normalize_to_uint8(np.mean(video_cleaned, axis=0))
+        ctx.set("M0_ff_image_cleaned", M0_ff_image_cleaned)
+
+        self.logger.info(f"    - Removed {len(arterial_pulse_filtered) - len(arterial_pulse_cleaned)} frames due to low correlation with median arterial pulse beat pattern")
+        ctx.output_manager.output("pulse_analysis", f"pre_arterial signal corrected", (arterial_pulse_cleaned, beat_signal), "signal", options={"multiple_signals": True, "legend": ["Original Signal", "beat signal"]})
 
         # --- Compute correlation map with filtered pulses ---
 
-        correlation_artery = signal_processing.compute_correlation(video_cleaned, arterial_pulse_interpolated)
-        # correlation_vein = pulse_analysis.compute_correlation(video, venous_pulse_filtered)
+        correlation_artery = signal_processing.compute_correlation(video_cleaned, arterial_pulse_cleaned)
         ctx.set("correlation", correlation_artery)
+        # correlation_vein = pulse_analysis.compute_correlation(video, venous_pulse_filtered)
         # ctx.set("correlation_vein", correlation_vein)
 
         # --- Accumulate frames at the systolic and diastolic peaks of the filtered pulses ---
 
-        diasys, M0_Systole_img, M0_Diastole_img = pulse_analysis.compute_diasys_image(video_cleaned, arterial_pulse_interpolated, sampling_frequency)
+        diasys, sysindexes, diasindexes = pulse_analysis.compute_diasys_image(video_cleaned, arterial_pulse_cleaned, sampling_frequency)
+        ctx.output_manager.output("pulse_analysis", f"diasys plot", (arterial_pulse_cleaned, sysindexes), "signal", options={"scatter": True})
 
-        ctx.set("pre_arterial_pulse", arterial_pulse)
-        ctx.set("pre_venous_pulse", venous_pulse)
-        ctx.set("choroidal_pulse", choroidal_pulse)
-        ctx.set("pre_arterial_pulse_filtered", arterial_pulse_filtered)
-        ctx.set("pre_venous_pulse_filtered", venous_pulse_filtered)
-        ctx.set("pre_arterial_pulse_interpolated", arterial_pulse_interpolated)
-        ctx.set("choroidal_pulse_filtered", choroidal_pulse_filtered)
         ctx.set("diasys_image", diasys)
+
+
+